@@ -1,11 +1,11 @@
 name: showcaseview
 description: A Flutter package to Showcase/Highlight widgets step by step.
-version: 0.1.8+18
+version: 0.2.0
 homepage: https://github.com/simformsolutions/flutter_showcaseview
 issue_tracker: https://github.com/simformsolutions/flutter_showcaseview/issues
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
